+++ helion/_compiler/pallas/backend.py
@@ -222,7 +222,7 @@ class PallasBackend(Backend):
         return f"{offsets_var} = {lid} * {block_size_var} + jnp.arange(0, {block_size_var}, dtype={dtype})"
 
     def sympy_printer_expr(self, expr: sympy.Expr) -> str:
-        from ..device_function import pallas_texpr
+        from .printer import pallas_texpr
 
         return pallas_texpr(expr)
